// src/app/core/models/customer.model.ts
export class Customer{
    customerId?: number;
    firstName?: string;
    lastName?: string;

    email?: string;

    businessName?: string;
    typeOfBusiness?: string;
    workPhone?: string;
    personalPhone?: string;

    businessStreetAddress?: string;
    businessCity?: string;
    businessState?: string;
    businessZip?: string;
    businessCountry?: string;

}